Backport CUMULUS-2387
