+++ packages/api/lambdas/process-s3-dead-letter-archive.js
@@ -5,6 +5,7 @@ const log = require('@cumulus/common/log');
 const { s3 } = require('@cumulus/aws-client/services');
 const { getJsonS3Object } = require('@cumulus/aws-client/S3');
 const { getKnexClient } = require('@cumulus/db');
+const { unwrapDeadLetterCumulusMessage } = require('@cumulus/message/DeadLetterMessage');
 
 const { writeRecords } = require('./sf-event-sqs-to-db-records');
 
@@ -43,7 +44,8 @@ async function processDeadLetterArchive({
     const deadLetterObjects = listObjectsResponse.Contents;
     const promises = await Promise.allSettled(deadLetterObjects.map(
       async (deadLetterObject) => {
-        const cumulusMessage = await getJsonS3Object(bucket, deadLetterObject.Key);
+        const deadLetterMessage = await getJsonS3Object(bucket, deadLetterObject.Key);
+        const cumulusMessage = await unwrapDeadLetterCumulusMessage(deadLetterMessage);
         try {
           await writeRecordsFunction({ cumulusMessage, knex });
           return deadLetterObject.Key;
