+++ packages/api/lambdas/sf-event-sqs-to-db-records/index.js
@@ -19,10 +19,11 @@ const {
 const {
   getMessageExecutionParentArn,
 } = require('@cumulus/message/Executions');
+const { getCumulusMessageFromExecutionEvent } = require('@cumulus/message/StepFunctions');
+
 const Execution = require('../../models/executions');
 const Granule = require('../../models/granules');
 const Pdr = require('../../models/pdrs');
-const { getCumulusMessageFromExecutionEvent } = require('../../lib/cwSfExecutionEventUtils');
 
 const {
   getCollectionCumulusId,
