+++ packages/api/tests/lambdas/test-write-db-dlq-records-to-s3.js
@@ -9,7 +9,6 @@ const { randomString } = require('@cumulus/common/test-utils');
 const {
   determineExecutionName,
   handler,
-  unwrapDeadLetterCumulusMessage,
 } = require('../../lambdas/write-db-dlq-records-to-s3.js');
 
 test.before(async (t) => {
@@ -120,84 +119,4 @@ test('determineExecutionName returns execution name if it exists in cumulus_meta
 
 test('determineExecutionName returns "unknown" if getExecutionName throws error', (t) => {
   t.is(determineExecutionName({}), 'unknown');
-});
-
-test('unwrapDeadLetterCumulusMessage unwraps an SQS message', (t) => {
-  const cumulusMessage = {
-    cumulus_meta: {
-      execution_name: randomString(),
-    },
-  };
-  const testMessage = {
-    eventSource: 'aws:sqs',
-    body: JSON.stringify(cumulusMessage),
-  };
-  t.deepEqual(unwrapDeadLetterCumulusMessage(testMessage), cumulusMessage);
-});
-
-test('unwrapDeadLetterCumulusMessage unwraps an AWS States message', (t) => {
-  const cumulusMessage = {
-    cumulus_meta: {
-      execution_name: randomString(),
-    },
-  };
-  const testMessage = {
-    source: 'aws.states',
-    detail: {
-      output: JSON.stringify(cumulusMessage),
-    },
-  };
-  t.deepEqual(unwrapDeadLetterCumulusMessage(testMessage), cumulusMessage);
-});
-
-test('unwrapDeadLetterCumulusMessage unwraps an AWS States message with only input', (t) => {
-  const cumulusMessage = {
-    cumulus_meta: {
-      execution_name: randomString(),
-    },
-  };
-  const testMessage = {
-    source: 'aws.states',
-    detail: {
-      input: JSON.stringify(cumulusMessage),
-    },
-  };
-  t.deepEqual(unwrapDeadLetterCumulusMessage(testMessage), cumulusMessage);
-});
-
-test('unwrapDeadLetterCumulusMessage unwraps an AWS states message within an SQS message', (t) => {
-  const cumulusMessage = {
-    cumulus_meta: {
-      execution_name: randomString(),
-    },
-  };
-  const testStatesMessage = {
-    source: 'aws.states',
-    detail: {
-      output: JSON.stringify(cumulusMessage),
-    },
-  };
-  const testSqsMessage = {
-    sourceEvent: 'aws:sqs',
-    body: JSON.stringify(testStatesMessage),
-  };
-  t.deepEqual(unwrapDeadLetterCumulusMessage(testSqsMessage), cumulusMessage);
-});
-
-test('unwrapDeadLetterCumulusMessage returns wrapped message on error', (t) => {
-  const invalidMessage = {
-    eventSource: 'aws:sqs',
-    detail: {},
-  };
-  t.deepEqual(unwrapDeadLetterCumulusMessage(invalidMessage), invalidMessage);
-});
-
-test('unwrapDeadLetterCumulusMessage returns an non-unwrappable message', (t) => {
-  const testMessage = {
-    eventSource: 'aws:something-strange',
-    contents: JSON.stringify({
-      key: 'value',
-    }),
-  };
-  t.deepEqual(unwrapDeadLetterCumulusMessage(testMessage), testMessage);
 });
